+++ src/main/scala/io/mkotsur/test/App.scala
@@ -1,13 +1,9 @@
 package io.mkotsur.test
 
-import cats.Semigroup
 import cats.effect._
 import io.circe.generic.auto._
-import org.http4s.{ EntityDecoder, _ }
-import org.http4s.client.Client
+import org.http4s._
 import org.http4s.client.blaze._
-import org.http4s.client.dsl.io._
-import org.http4s.headers._
 import org.http4s.util.CaseInsensitiveString
 
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -16,6 +12,7 @@ import scala.io.StdIn
 object App extends IOApp {
 
   import Decoders.implicits._
+  import GithubApi._
   import Output.implicits._
   import cats.implicits._
 
@@ -23,21 +20,7 @@ object App extends IOApp {
     case None    => IO.raiseError(new RuntimeException(s"Please set $envKey env variable"))
     case Some(t) => IO.pure(Credentials.Token(CaseInsensitiveString("token"), t))
   }
-  private def getRepoInfosUri(username: String) = IO(Uri.uri("https://api.github.com/users/") / username / "repos")
-
-  private def getAllPages[A: Semigroup](
-    uri: Uri,
-    token: Credentials.Token
-  )(implicit decoder: EntityDecoder[IO, (A, Option[Uri])], client: Client[IO]): IO[A] = {
-    val req = Method.GET(uri, Authorization(token))
-    client.expect[(A, Option[Uri])](req).flatMap {
-      case (items, None) => IO.pure(items)
-      case (items, Some(nextPageUri)) =>
-        getAllPages[A](nextPageUri, token).map { newItems =>
-          newItems |+| items
-        }
-    }
-  }
+  private def makeRepoInfosUri(username: String) = IO(Uri.uri("https://api.github.com/users/") / username / "repos")
 
   case class Contributor(login: String, contributions: Int)
 
@@ -49,7 +32,7 @@ object App extends IOApp {
         token        <- readGithubToken("GH_TOKEN")
         _            <- IO(println("Please enter your Github ID ?"))
         username     <- IO(StdIn.readLine())
-        repoInfosUri <- getRepoInfosUri(username)
+        repoInfosUri <- makeRepoInfosUri(username)
         repoInfos    <- getAllPages[List[RepoInfo]](repoInfosUri, token)
         _            <- IO(println(s"Found ${repoInfos.length} repos"))
         contributorsPerRepo <- repoInfos
